+++ compat/gray.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stdint.h>
 #define GRAYDBUFFER_SIZE 7688
 
 #define DEFAULT_FPS 30
@@ -9,11 +10,11 @@ enum GrayPlanes {LIGHT_PLANE = 0, DARK_PLANE = 1};
 #define GrayOn() (1)
 #define GrayOff() do {} while (0)
 
-void DelayNFrames(unsigned int frames);
+void DelayNFrames(uint16_t frames);
 
 void GrayDBufInit(void *buf);
-void *GrayDBufGetActivePlane(short plane);
-void *GrayDBufGetHiddenPlane(short plane);
+void *GrayDBufGetActivePlane(int16_t plane);
+void *GrayDBufGetHiddenPlane(int16_t plane);
 
 void GrayDBufToggleSync(void);
 void GrayDBufRefresh(void);
